Fix node, property, and label parsing
Added "_" to the list of characters allowed for labels.

Added "?" to the list of characters allowed for property names.

Changed node names to allow starting with any of the allowed characters.
The specification states that a node name shall start with a lower or
uppercase letter, however Zephyr's parser does not enforce this.

Fixes #20